requirements.txt
# Web Framework
Flask==2.3.3
Flask-CORS==4.0.0
Flask-Migrate==4.0.5

# Database
psycopg2-binary==2.9.7
SQLAlchemy==2.0.23
Flask-SQLAlchemy==3.1.1

# Authentication & Security
bcrypt==4.0.1
PyJWT==2.8.0
Flask-Login==0.6.3
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.5.3

# Email
Flask-Mail==0.9.1

# Environment Variables
python-dotenv==1.0.0

# Production Server
gunicorn==21.2.0

# Utilities
Werkzeug==2.3.7
requests==2.31.0
python-multipart==0.0.6

# Development
python-dotenv==1.0.0